requirements.txt
groq
datasets
rouge-score
bert-score
reportlab
matplotlib
pandas
huggingface_hub
python-dotenv
numpy
